#7 Moved sql store to subfolder
